clean and refactor code
